+++ graal/com.oracle.graal.graph/src/com/oracle/graal/graph/Node.java
@@ -136,8 +136,6 @@ public abstract class Node implements Cloneable, Formattable {
     private static final int INLINE_USAGE_COUNT = 2;
     private static final Node[] NO_NODES = {};
 
-    private static final boolean VERIFY_TYPES = true;
-
     /**
      * Head of usage list. The elements of the usage list in order are {@link #usage0},
      * {@link #usage1} and {@link #extraUsages}. The first null entry terminates the list.
@@ -748,14 +746,12 @@ public abstract class Node implements Cloneable, Formattable {
             for (Node usage : usages()) {
                 assertFalse(usage.isDeleted(), "usage %s must never be deleted", usage);
                 assertTrue(usage.inputs().contains(this), "missing input in usage %s", usage);
-                if (VERIFY_TYPES) {
-                    NodeClassIterator iterator = usage.inputs().iterator();
-                    while (iterator.hasNext()) {
-                        Position pos = iterator.nextPosition();
-                        if (pos.get(usage) == this && pos.getInputType(usage) != InputType.Unchecked) {
-                            assert isAllowedUsageType(pos.getInputType(usage)) : "invalid input of type " + pos.getInputType(usage) + " from " + usage + " to " + this + " (" +
-                                            pos.getInputName(usage) + ")";
-                        }
+                NodeClassIterator iterator = usage.inputs().iterator();
+                while (iterator.hasNext()) {
+                    Position pos = iterator.nextPosition();
+                    if (pos.get(usage) == this && pos.getInputType(usage) != InputType.Unchecked) {
+                        assert isAllowedUsageType(pos.getInputType(usage)) : "invalid input of type " + pos.getInputType(usage) + " from " + usage + " to " + this + " (" + pos.getInputName(usage) +
+                                        ")";
                     }
                 }
             }