+++ graal/com.oracle.graal.nodes/src/com/oracle/graal/nodes/GuardPhiNode.java
@@ -26,7 +26,7 @@ import com.oracle.graal.graph.*;
 import com.oracle.graal.nodes.extended.*;
 import com.oracle.graal.nodes.type.*;
 
-@NodeInfo(nameTemplate = "GuardPhi({i#values})")
+@NodeInfo(nameTemplate = "GuardPhi({i#values})", allowedUsageTypes = {InputType.Guard})
 public class GuardPhiNode extends PhiNode implements GuardingNode {
 
     @Input(InputType.Guard) final NodeInputList<ValueNode> values = new NodeInputList<>(this);